// src/app/model/company/company.ts
export class CompanyModel {
    company_id: number;
    company_title: string;
    company_url: string;
    company_address: string;
    company_phone: string;
    company_email: string;
    company_hr_phone: string;
    company_hr_emailid: string;
    smtp_host: string;
    smtp_port: number;
    smtp_username: string;
    smtp_password: string;
    created_by: string;
    updated_by: string;
    created_datetime: string;
    updated_datetime: string;
}